+++ src/components/InfographicEditor/Editors/Components/DropdownList.js
@@ -2,11 +2,16 @@ import React from 'react';
 
 class DropdownList extends React.Component
 {
+    static defaultProps = {
+        isDisabled: false,
+    }
+
     constructor(props)
     {
         super(props);
         this.state = {
             value: this.props.selected,
+            isDisabled: this.props.isDisabled,
         }
         this._handleChange = this._HandleChange.bind(this);
     }
@@ -15,7 +20,8 @@ class DropdownList extends React.Component
     {
         return (
             <div>
-                <select value={this.state.value} onChange={this._handleChange}>
+                <select value={this.state.value} onChange={this._handleChange}
+                    disabled={this.state.isDisabled}>
                     {
                         this.props.options.map(d => {
                             return (<option value={d}>{d}</option>)
@@ -30,6 +36,10 @@ class DropdownList extends React.Component
     {
         if (prevProps.selected !== this.props.selected) {
             this.setState({ value: this.props.selected });
+        }
+
+        if (prevProps.isDisabled !== this.props.isDisabled) {
+            this.setState({ isDisabled: this.props.isDisabled });
         }
     }